@holiman 2019-12-31
This is now up for a benchmark full sync: https://geth-bench.ethdevops.io/d/Jpk-Be5Wk/dual-geth?orgId=1&var-exp=mon08&var-master=mon09&var-percentile=50&from=1577792859000&to=now