@holiman 2020-01-01
Some preliminary charts. So, this PR only does paralell hashing if the number of updates is 100 or more, so as you can see, it has basically no effect pre-byzantium (when we hash in between every tx). However, the effect post-byz is _very_ visible: 

![Screenshot_2020-01-01 Dual Geth - Grafana](https://user-images.githubusercontent.com/142290/71639765-18fd2980-2c7d-11ea-8685-143537767994.png)

Here's a bit more zoom = this PR is pretty stable at about `1.9ms`, and master varies between `2.5 - 5 ms`

![Screenshot_2020-01-01 Dual Geth - Grafana(1)](https://user-images.githubusercontent.com/142290/71639777-61b4e280-2c7d-11ea-8ceb-1e04e09e7440.png)