@holiman 2020-01-01
Oh wow - `account hash` has now moved down from number 2: 
![Screenshot_2020-01-01 Dual Geth - Grafana(2)](https://user-images.githubusercontent.com/142290/71642243-95f1c880-2ca8-11ea-9631-c813a3de4e5e.png)