@holiman 2020-01-02
These are the first `5.8M` blocks (bottom is this PR). 
![Screenshot_2020-01-02 Dual Geth - Grafana](https://user-images.githubusercontent.com/142290/71684559-25c17080-2d96-11ea-9245-5f7a9887bd8e.png)

In total, `account hash` has dropped to seventh place, from fourth. 
Here's the `account hash` individually
![Screenshot_2020-01-02 Dual Geth - Grafana(1)](https://user-images.githubusercontent.com/142290/71684650-5903ff80-2d96-11ea-8174-c94d43b59c6e.png)