@holiman 2020-01-05
After ~5 days, `118h`, it's `10h 40m` ahead.