@holiman 2020-01-05
The other one (https://github.com/ethereum/go-ethereum/pull/20493) , running on `06/07`, is roughly `17h30m` ahead after `116h`. I don't think it's because that one is so much faster than this one, but merely that `06` (master) is a lot slower than `07` (experimental). The true number for the improvement is probably somewhere in between.