@holiman 2020-01-06
Let's do https://github.com/ethereum/go-ethereum/pull/20481 first, and once that one is 'done', I'll rebase this and fix the hasher-specific things