@holiman 2020-01-07
When this was executed without the paralell hashing, this was the result:

> Mon08 (this PR) finished full sync at block 9188701 at 2019-12-31 03:05:02, after a total of 175h
mon09 was 10 hours behind, at block 9.177M, when the run was stopped.

With the addition of parallel hashing, the new results are, it hit `9188754` at `Jan 07 13:45:20`, starting at `Dec-31, 12:47`. So roughly `7 x 24-1 hours`= `167h`, shaving off another `8` hours from https://github.com/ethereum/go-ethereum/pull/20481. It looks to be somewhere around `16h` ahead of `master` on `mon09`.